fixed issue with using authorInfo projecttion for AUthorRepository, both projections were moved to the subpackage of domain type
