+++ src/main/java/com/mogikanensoftware/spring/demo/app/entity/projection/AuthorInfo.java
@@ -1,4 +1,4 @@
-package com.mogikanensoftware.spring.demo.app.projections;
+package com.mogikanensoftware.spring.demo.app.entity.projection;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.data.rest.core.config.Projection;
@@ -8,6 +8,7 @@ import com.mogikanensoftware.spring.demo.app.entity.Author;
 @Projection(name = "authorInfo", types = Author.class)
 public interface AuthorInfo {
 
+	@Value("#{target.getId()}")
 	Long getId();
 
 	@Value("#{target.getFirstName()}")
