+++ src/main/java/com/mogikanensoftware/spring/demo/app/repo/AuthorRepository.java
@@ -5,12 +5,11 @@ import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 import org.springframework.security.access.prepost.PreAuthorize;
 
 import com.mogikanensoftware.spring.demo.app.entity.Author;
-import com.mogikanensoftware.spring.demo.app.projections.AuthorInfo;
+import com.mogikanensoftware.spring.demo.app.entity.projection.AuthorInfo;
 
-//@RepositoryRestResource(exported=true, excerptProjection = AuthorInfo.class)
-@RepositoryRestResource(exported=true)
+@RepositoryRestResource(exported = true, excerptProjection = AuthorInfo.class)
 @PreAuthorize("hasRole('VIEW')")
-public interface AuthorRepository extends PagingAndSortingRepository<Author, Long>{
+public interface AuthorRepository extends PagingAndSortingRepository<Author, Long> {
 
 	@PreAuthorize("hasRole('DELETE')")
 	@Override
