+++ src/main/java/com/mogikanensoftware/spring/demo/app/repo/BookRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 import org.springframework.security.access.prepost.PreAuthorize;
 
 import com.mogikanensoftware.spring.demo.app.entity.Book;
-import com.mogikanensoftware.spring.demo.app.projections.FullBookInfo;
+import com.mogikanensoftware.spring.demo.app.entity.projection.FullBookInfo;
 
 @RepositoryRestResource(exported = true, excerptProjection=FullBookInfo.class)
 @PreAuthorize("hasRole('VIEW')")
