+++ src/test/resources/api/responses/get/authors.json
@@ -1,0 +1,153 @@
+{
+  "_embedded": {
+    "authors": [
+      {
+        "firstName": "Yakov",
+        "lastName": "Fain",
+        "_links": {
+          "self": {
+            "href": "http://localhost:8080/api/v1/authors/1"
+          },
+          "author": {
+            "href": "http://localhost:8080/api/v1/authors/1"
+          }
+        }
+      },
+      {
+        "firstName": "Brian",
+        "lastName": "Goetz",
+        "_links": {
+          "self": {
+            "href": "http://localhost:8080/api/v1/authors/2"
+          },
+          "author": {
+            "href": "http://localhost:8080/api/v1/authors/2"
+          }
+        }
+      },
+      {
+        "firstName": "Martin",
+        "lastName": "Fowler",
+        "_links": {
+          "self": {
+            "href": "http://localhost:8080/api/v1/authors/3"
+          },
+          "author": {
+            "href": "http://localhost:8080/api/v1/authors/3"
+          }
+        }
+      },
+      {
+        "firstName": "Sam",
+        "lastName": "Newman",
+        "_links": {
+          "self": {
+            "href": "http://localhost:8080/api/v1/authors/4"
+          },
+          "author": {
+            "href": "http://localhost:8080/api/v1/authors/4"
+          }
+        }
+      },
+      {
+        "firstName": "Tom",
+        "lastName": "DeMarco",
+        "_links": {
+          "self": {
+            "href": "http://localhost:8080/api/v1/authors/5"
+          },
+          "author": {
+            "href": "http://localhost:8080/api/v1/authors/5"
+          }
+        }
+      },
+      {
+        "firstName": "Rober",
+        "lastName": "C. Martin",
+        "_links": {
+          "self": {
+            "href": "http://localhost:8080/api/v1/authors/6"
+          },
+          "author": {
+            "href": "http://localhost:8080/api/v1/authors/6"
+          }
+        }
+      },
+      {
+        "firstName": "Rajesh",
+        "lastName": "RV",
+        "_links": {
+          "self": {
+            "href": "http://localhost:8080/api/v1/authors/7"
+          },
+          "author": {
+            "href": "http://localhost:8080/api/v1/authors/7"
+          }
+        }
+      },
+      {
+        "firstName": "Noel",
+        "lastName": "Markham",
+        "_links": {
+          "self": {
+            "href": "http://localhost:8080/api/v1/authors/8"
+          },
+          "author": {
+            "href": "http://localhost:8080/api/v1/authors/8"
+          }
+        }
+      },
+      {
+        "firstName": "Tomek",
+        "lastName": "Kaczanowski",
+        "_links": {
+          "self": {
+            "href": "http://localhost:8080/api/v1/authors/9"
+          },
+          "author": {
+            "href": "http://localhost:8080/api/v1/authors/9"
+          }
+        }
+      },
+      {
+        "firstName": "Walter",
+        "lastName": "Isaacson",
+        "_links": {
+          "self": {
+            "href": "http://localhost:8080/api/v1/authors/10"
+          },
+          "author": {
+            "href": "http://localhost:8080/api/v1/authors/10"
+          }
+        }
+      },
+      {
+        "firstName": "Stoyan",
+        "lastName": "Stefanov",
+        "_links": {
+          "self": {
+            "href": "http://localhost:8080/api/v1/authors/11"
+          },
+          "author": {
+            "href": "http://localhost:8080/api/v1/authors/11"
+          }
+        }
+      }
+    ]
+  },
+  "_links": {
+    "self": {
+      "href": "http://localhost:8080/api/v1/authors{?page,size,sort}",
+      "templated": true
+    },
+    "profile": {
+      "href": "http://localhost:8080/api/v1/profile/authors"
+    }
+  },
+  "page": {
+    "size": 20,
+    "totalElements": 11,
+    "totalPages": 1,
+    "number": 0
+  }
+}
